Flags package refactor to vortexengine.flags
